rag-agent/requirements.txt
azure-core==1.36.0
azure-cosmos==4.14.2
azure-functions==1.24.0
azure-functions-durable==1.4.0
azure-identity==1.25.1
openai==2.9.0
openai-agents==0.6.2
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0